Split export.py along its own seams, and drop two dead writers
830 lines in one module, already sectioned by banner comments. Those sections become files:

  export/graph.py     opset and IR pins, saving, renaming, topological order
  export/isotonic.py  the calibrator as a graph, thinned to a workable knot count
  export/gate.py      T and its ramp
  export/heads.py     the blender, the probability picker, the soft model's proto
  export/metadata.py  what the bundle says about itself, and the names it is known by
  export/parity.py    the probe and tolerance that decide whether a fuse may ship
  export/fuse.py      the column plan and the stitch
  export/bundle.py    build_bundle
  export/__init__.py  the module docstring, and build_bundle

Bodies moved verbatim by line range. `build_bundle` is the only entry point any caller uses, so
cli.py and tests/conftest.py are untouched.

`isotonic_to_onnx` and `tanimoto_to_onnx` are deleted. Both wrote one stage to its own file and
self-checked it against the Python original — useful before `build_bundle` learned to check the whole
fused graph the same way, and callerless ever since. AST analysis over the tree found only their own
definitions: no string dispatch, no registry, no getattr. Their helpers stay, since `_fuse` builds
both stages inline. `_rename_input` looked dead too and is not — `_tanimoto_model` calls it.
