--- olinda/export.py
@@ -1,919 +1,0 @@
-"""Fuse a trained model dir into a single, self-describing ``model.onnx``.
-
-Every olinda-owned transform downstream of the RDKit featurizer is ONNX-able, so a bundle collapses to ONE
-graph that runs on onnxruntime alone. Two shapes:
-
-- **soft-only**: ``fp → soft_model → [soft_correction] → prediction`` (= ``S``).
-- **hard present**: fuse ``S``, ``H → h_correction`` (``H_S``), ``T → ramp`` (the weight ``a``) and the
-  blender, ``(1-a)·S + a·H_S``.
-
-Either way a column declares **one output, its prediction**, named after the teacher column. ``S``,
-``H_S`` and ``a`` stay internal tensors: they are how the answer is computed, not part of the answer,
-and declaring them would let callers depend on a wiring that has already changed once.
-
-The **featurizer config + provenance travel inside ``model.onnx`` metadata** (``metadata_props["olinda"]``),
-so the file is self-describing — a consumer reads the Morgan config (and RDKit version) to build the 2048-count
-fingerprint in Python (no ONNX op for featurization) and runs the single graph.
-
-The hard head is task-aware: a **classifier** exposes ``probabilities`` (we take column 1); a **regressor**
-would expose ``variable`` directly (seam — only classifier is enabled today, see :func:`olinda.hard`).
-"""
-
-from __future__ import annotations
-
-import json
-import tempfile
-from datetime import datetime, timezone
-from pathlib import Path
-
-import numpy as np
-
-from olinda.console import STEP_COLORS, echo, rule, success, summary_panel
-from olinda.console import path as cpath
-from olinda.hard.layout import (
-    H_TO_S_NAME,
-    HARD_DIRNAME,
-    HARD_EVAL_NAME,
-    HARD_H5_NAME,
-    HARD_META_NAME,
-    HARD_MODEL_SUBDIR,
-    TANIMOTO_DIRNAME,
-    has_hard_head,
-)
-from olinda.metrics import json_safe
-
-MODEL_NAME = "model.onnx"
-# One schema, versioned from here on. Nothing in the wild predates it, so there is no legacy shape to
-# read: OlindaArtifact refuses a bundle it does not recognise and says to re-run `olinda export`.
-BUNDLE_SCHEMA = "olinda.bundle.v1"
-PRODUCER_NAME = "olinda"
-_IR_VERSION = 10  # onnxruntime in this env caps the model IR version at 10
-_OPSET = 16
-_ISOTONIC_TOL = 1e-5  # knot-thinning target — kept below _PARITY_TOL so ONNX float error has headroom
-_ISOTONIC_MAX_KNOTS = 4096
-_PARITY_TOL = 1e-4  # build_bundle raises if the fused graph drifts from the Python reference beyond this
-_SAMPLE_SMILES = ["CCO", "c1ccccc1", "CC(=O)O", "CCN", "OCc1ccccc1"]
-_PARITY_LABELLED = (
-    8  # labelled compounds per hard column added to the probe, to exercise the blend
-)
-
-
-_PROTOBUF_LIMIT = 2 * 1024**3  # a single serialized ModelProto cannot exceed this
-
-
-def _strip_dead_attributes(model) -> int:
-    """Drop tree attributes that carry no information; returns the bytes saved.
-
-    ``nodes_hitrates`` is emitted as one float per node and, for boosters converted from LightGBM or
-    XGBoost, is uniformly ``1.0`` — on a real model that is 1.5M copies of the same constant, about an
-    eighth of the file, which onnxruntime ignores. It is optional in the ai.onnx.ml schema, so removing
-    it leaves predictions bit-identical.
-
-    ``nodes_missing_value_tracks_true`` looks similar but genuinely carries both values, so it stays.
-    """
-    import numpy as np
-
-    saved = 0
-    for node in model.graph.node:
-        if not node.op_type.startswith("TreeEnsemble"):
-            continue
-        keep = []
-        for attr in node.attribute:
-            if (
-                attr.name == "nodes_hitrates"
-                and len(attr.floats)
-                and np.all(np.asarray(attr.floats) == 1.0)
-            ):
-                saved += attr.ByteSize()
-                continue
-            keep.append(attr)
-        if len(keep) != len(node.attribute):
-            del node.attribute[:]
-            node.attribute.extend(keep)
-    return saved
-
-
-def _save(model, path: Path) -> None:
-    import onnx
-
-    model.ir_version = _IR_VERSION
-    _strip_dead_attributes(model)
-    size = model.ByteSize()
-    if size >= _PROTOBUF_LIMIT:
-        raise ValueError(
-            f"fused model is {size / 1e9:.2f} GB, over ONNX's {_PROTOBUF_LIMIT / 1e9:.2f} GB protobuf limit. "
-            "Reduce the number of columns or --num-boost-round."
-        )
-    onnx.checker.check_model(model)
-    onnx.save(model, str(path))
-
-
-# ── isotonic calibrator → ONNX ────────────────────────────────────────────────
-
-
-def _rdp_vertical(x: np.ndarray, y: np.ndarray, tol: float) -> np.ndarray:
-    """Ramer–Douglas–Peucker on the ``(x, y)`` polyline with a VERTICAL tolerance; return the kept-point mask.
-
-    Bounding the deviation at the anchors bounds it everywhere (both maps are piecewise-linear over the
-    anchors) — unlike grid sampling, which only controls the sampled points.
-    """
-    n = len(x)
-    keep = np.zeros(n, dtype=bool)
-    keep[0] = keep[-1] = True
-    stack = [(0, n - 1)]
-    while stack:
-        i, j = stack.pop()
-        if j <= i + 1:
-            continue
-        xs = x[i : j + 1]
-        dx = x[j] - x[i]
-        chord = (
-            y[i] + (y[j] - y[i]) * ((xs - x[i]) / dx)
-            if dx != 0
-            else np.full(len(xs), y[i])
-        )
-        d = np.abs(y[i : j + 1] - chord)
-        k = int(d.argmax())
-        if d[k] > tol:
-            keep[i + k] = True
-            stack.append((i, i + k))
-            stack.append((i + k, j))
-    return keep
-
-
-def _thin_isotonic(x: np.ndarray, y: np.ndarray, tol: float, max_knots: int):
-    """Simplify the monotone map to within ``tol`` via vertical-distance RDP (knots are true anchors)."""
-    if len(x) <= 2:
-        return x, y
-    t = tol
-    for _ in range(40):
-        keep = _rdp_vertical(x, y, t)
-        if int(keep.sum()) <= max_knots:
-            break
-        t *= 2
-    return x[keep], y[keep]
-
-
-def _isotonic_model(cal, in_name: str, out_name: str):
-    """ONNX ``ModelProto``: raw(float32) → ``np.interp(sign*raw, x, y)`` in float64 (searchsorted + one step)."""
-    from onnx import TensorProto, helper
-
-    xk, yk = _thin_isotonic(cal._x, cal._y, _ISOTONIC_TOL, _ISOTONIC_MAX_KNOTS)
-    xk = xk.astype(np.float64)
-    yk = yk.astype(np.float64)
-    slope = np.zeros_like(xk)
-    slope[:-1] = (yk[1:] - yk[:-1]) / (xk[1:] - xk[:-1])  # slope[-1]=0 → end-clamp
-    n = len(xk)
-    ct = helper.make_tensor
-    init = [
-        ct("sign", TensorProto.DOUBLE, [1], [float(cal._sign)]),
-        ct("xk", TensorProto.DOUBLE, [n], xk),
-        ct("yk", TensorProto.DOUBLE, [n], yk),
-        ct("slope", TensorProto.DOUBLE, [n], slope),
-        ct("x0", TensorProto.DOUBLE, [1], [float(xk[0])]),
-        ct("xlast", TensorProto.DOUBLE, [1], [float(xk[-1])]),
-        ct("one_i", TensorProto.INT64, [1], [1]),
-        ct("nm1", TensorProto.INT64, [1], [n - 1]),
-        ct("zero_i", TensorProto.INT64, [1], [0]),
-        ct("axis1", TensorProto.INT64, [1], [1]),
-    ]
-    nodes = [
-        helper.make_node("Cast", [in_name], ["rawd"], to=TensorProto.DOUBLE),
-        helper.make_node("Mul", ["rawd", "sign"], ["t_raw"]),
-        helper.make_node(
-            "Max", ["t_raw", "x0"], ["t_lo"]
-        ),  # clamp into [x0, xlast] → out-of-range holds ends
-        helper.make_node("Min", ["t_lo", "xlast"], ["t"]),
-        helper.make_node("Unsqueeze", ["t", "axis1"], ["t2"]),
-        helper.make_node("LessOrEqual", ["xk", "t2"], ["le"]),
-        helper.make_node("Cast", ["le"], ["lei"], to=TensorProto.INT64),
-        helper.make_node("ReduceSum", ["lei", "axis1"], ["cnt"], keepdims=0),
-        helper.make_node("Sub", ["cnt", "one_i"], ["idx0"]),
-        helper.make_node("Max", ["idx0", "zero_i"], ["idxc"]),
-        helper.make_node("Min", ["idxc", "nm1"], ["idx"]),
-        helper.make_node("Gather", ["xk", "idx"], ["x_i"]),
-        helper.make_node("Gather", ["yk", "idx"], ["y_i"]),
-        helper.make_node("Gather", ["slope", "idx"], ["m_i"]),
-        helper.make_node("Sub", ["t", "x_i"], ["dx"]),
-        helper.make_node("Mul", ["dx", "m_i"], ["step"]),
-        helper.make_node("Add", ["y_i", "step"], [out_name]),
-    ]
-    g = helper.make_graph(
-        nodes,
-        "isotonic",
-        [helper.make_tensor_value_info(in_name, TensorProto.FLOAT, ["B"])],
-        [helper.make_tensor_value_info(out_name, TensorProto.DOUBLE, ["B"])],
-        initializer=init,
-    )
-    return helper.make_model(g, opset_imports=[helper.make_opsetid("", _OPSET)])
-
-
-def isotonic_to_onnx(
-    cal, out_path: Path, *, in_name: str = "input", out_name: str = "output"
-) -> dict:
-    """Save a standalone isotonic ONNX and self-check vs ``cal.transform``; return ``{knots, max_abs_diff}``."""
-    import onnxruntime as ort
-
-    model = _isotonic_model(cal, in_name, out_name)
-    _save(model, out_path)
-    span = float(cal._x.max() - cal._x.min()) or 1.0
-    grid = np.linspace(
-        float(cal._x.min()) - 0.1 * span, float(cal._x.max()) + 0.1 * span, 2000
-    )
-    raw = (cal._sign * grid).astype(np.float32)
-    sess = ort.InferenceSession(
-        out_path.read_bytes(), providers=["CPUExecutionProvider"]
-    )
-    got = sess.run(None, {in_name: raw})[0]
-    n_knots = int(
-        len(_thin_isotonic(cal._x, cal._y, _ISOTONIC_TOL, _ISOTONIC_MAX_KNOTS)[0])
-    )
-    return {
-        "knots": n_knots,
-        "max_abs_diff": float(np.max(np.abs(got - cal.transform(raw)))),
-    }
-
-
-# ── T (Tanimoto regressor) + ramp → ONNX ──────────────────────────────────────
-
-
-def _tanimoto_model(clf, n_features: int, in_name: str, out_name: str):
-    """ONNX ``ModelProto``: fp(float32) → blend weight ``a`` (double).
-
-    Three parts spliced into one graph: binarise the incoming count fingerprint, run the gate network,
-    ramp its output into a weight.
-
-    The binarisation is not optional. The net is trained and evaluated on ``bits > 0``
-    (:meth:`TanimotoRegressor.predict_tanimoto`), while the fused graph carries the shared *count*
-    fingerprint, so without it the graph would feed counts to a net that has only ever seen indicators
-    and quietly disagree with its own Python reference on any molecule with a repeated substructure.
-
-    The ramp clips to [0, 1] before scaling: a net can extrapolate past the target's range, and a
-    similarity above 1 would otherwise push ``a`` past its ceiling.
-    """
-    import onnx
-    from onnx import TensorProto, helper
-
-    net = onnx.load_from_string(clf.onnx_bytes)
-
-    # Rewire the net's own input so the binarisation can sit in front of it.
-    inner = "t_bits"
-    _rename_input(net, inner)
-    del net.graph.input[:]
-    net.graph.input.append(
-        helper.make_tensor_value_info(
-            in_name, TensorProto.FLOAT, ["B", int(n_features)]
-        )
-    )
-
-    net_out = net.graph.output[0].name
-    span = max(float(clf.sim_hi) - float(clf.sim_lo), 1e-9)
-    ct = helper.make_tensor
-    net.graph.initializer.extend(
-        [
-            ct("t_thr", TensorProto.FLOAT, [1], [0.0]),
-            ct("t_zero", TensorProto.DOUBLE, [1], [0.0]),
-            ct("t_one", TensorProto.DOUBLE, [1], [1.0]),
-            ct("t_lo", TensorProto.DOUBLE, [1], [float(clf.sim_lo)]),
-            ct("t_span", TensorProto.DOUBLE, [1], [span]),
-            ct("t_max", TensorProto.DOUBLE, [1], [float(clf.a_max)]),
-            ct("t_flat", TensorProto.INT64, [1], [-1]),
-        ]
-    )
-    head = [
-        helper.make_node("Greater", [in_name, "t_thr"], ["t_on"]),
-        helper.make_node("Cast", ["t_on"], [inner], to=TensorProto.FLOAT),
-    ]
-    tail = [
-        helper.make_node("Cast", [net_out], ["t_simd"], to=TensorProto.DOUBLE),
-        helper.make_node("Reshape", ["t_simd", "t_flat"], ["t_sim1"]),
-        helper.make_node("Clip", ["t_sim1", "t_zero", "t_one"], ["t_sim"]),
-        helper.make_node("Sub", ["t_sim", "t_lo"], ["t_shift"]),
-        helper.make_node("Div", ["t_shift", "t_span"], ["t_frac0"]),
-        helper.make_node("Clip", ["t_frac0", "t_zero", "t_one"], ["t_frac"]),
-        helper.make_node("Mul", ["t_frac", "t_max"], [out_name]),
-    ]
-    existing = list(net.graph.node)
-    del net.graph.node[:]
-    net.graph.node.extend(head + existing + tail)
-
-    del net.graph.output[:]
-    net.graph.output.append(
-        helper.make_tensor_value_info(out_name, TensorProto.DOUBLE, ["B"])
-    )
-    # The gate pins its own opset when it serialises itself; re-stamp it to the bundle's. Every op used
-    # here (MatMul, Relu, Add, Cast, Reshape, Clip, Greater) exists at _OPSET, and unlike the boosted
-    # stages there is no ai.onnx.ml domain to preserve, so replacing the whole list is safe.
-    del net.opset_import[:]
-    net.opset_import.append(helper.make_opsetid("", _OPSET))
-    net.ir_version = _IR_VERSION
-    return net
-
-
-def _rename_input(model, new_name: str) -> None:
-    """Point a converted booster's graph at ``new_name`` so the fuse can wire it up like any other stage."""
-    old = model.graph.input[0].name
-    model.graph.input[0].name = new_name
-    for node in model.graph.node:
-        node.input[:] = [new_name if i == old else i for i in node.input]
-
-
-def tanimoto_to_onnx(
-    clf, n_features: int, out_path: Path, *, in_name: str = "input"
-) -> dict:
-    """Save ``T`` + its ramp as a standalone ONNX and self-check vs ``clf.weight``; return ``{max_abs_diff}``."""
-    import onnxruntime as ort
-
-    _save(_tanimoto_model(clf, n_features, in_name, "a"), out_path)
-    rng = np.random.default_rng(0)
-    # Morgan *counts*, not bits: a 0/1 probe cannot tell "the graph binarises" from "the graph forgot
-    # to", because the two agree exactly on such input. Real fingerprints reach well past 1.
-    fp = (rng.random((64, n_features)) < 0.1) * rng.integers(1, 5, (64, n_features))
-    fp = fp.astype(np.float32)
-    sess = ort.InferenceSession(
-        out_path.read_bytes(), providers=["CPUExecutionProvider"]
-    )
-    got = sess.run(None, {in_name: fp})[0]
-    return {"max_abs_diff": float(np.max(np.abs(got - clf.weight(fp))))}
-
-
-# ── small structural graphs (blender, prob→H) ─────────────────────────────────
-
-
-def _blender_model():
-    """ONNX ``ModelProto``: (soft, hard, a) → ``(1-a)*soft + a*hard`` (all double)."""
-    from onnx import TensorProto, helper
-
-    vi = lambda n: helper.make_tensor_value_info(n, TensorProto.DOUBLE, ["B"])  # noqa: E731
-    nodes = [
-        helper.make_node("Sub", ["one", "a"], ["oma"]),
-        helper.make_node("Mul", ["oma", "soft"], ["p0"]),
-        helper.make_node("Mul", ["a", "hard"], ["p1"]),
-        helper.make_node("Add", ["p0", "p1"], ["prediction"]),
-    ]
-    g = helper.make_graph(
-        nodes,
-        "blender",
-        [vi("soft"), vi("hard"), vi("a")],
-        [vi("prediction")],
-        initializer=[helper.make_tensor("one", TensorProto.DOUBLE, [1], [1.0])],
-    )
-    return helper.make_model(g, opset_imports=[helper.make_opsetid("", _OPSET)])
-
-
-def _prob1_model(in_name: str, out_name: str):
-    """ONNX ``ModelProto``: classifier ``probabilities`` (B,2) → column 1 (B,) (a classifier's positive score)."""
-    from onnx import TensorProto, helper
-
-    nodes = [helper.make_node("Gather", [in_name, "one_idx"], [out_name], axis=1)]
-    g = helper.make_graph(
-        nodes,
-        "prob1",
-        [helper.make_tensor_value_info(in_name, TensorProto.FLOAT, ["B", 2])],
-        [helper.make_tensor_value_info(out_name, TensorProto.FLOAT, ["B"])],
-        initializer=[
-            helper.make_tensor("one_idx", TensorProto.INT64, [], [1])
-        ],  # scalar → drops axis 1
-    )
-    return helper.make_model(g, opset_imports=[helper.make_opsetid("", _OPSET)])
-
-
-def _soft_model_proto(sm, x_dim: int):
-    """The surrogate regressor as an ONNX ``ModelProto`` (via the backend exporter, through a temp file)."""
-    import onnx
-
-    from olinda.train.backend import get_backend
-
-    with tempfile.TemporaryDirectory() as td:
-        p = Path(td) / "s.onnx"
-        get_backend(sm.backend, "cpu").to_onnx(sm.model, p, x_dim)
-        return onnx.load(str(p))
-
-
-# ── fusion ────────────────────────────────────────────────────────────────────
-
-
-def _task(kind: str, value_range=None) -> dict:
-    """One task descriptor: what a head or a column predicts, and over what range.
-
-    Every predicting thing in the bundle carries one of these, so nothing has to be inferred from which
-    metric keys happen to be present. ``kind`` is ``"regression"`` or ``"classification"``; ``range`` is
-    the observed span of the values, or ``None`` when it was not measured. ``units`` is reserved and
-    always ``None`` today — olinda never learns what a teacher's numbers mean.
-    """
-    return {
-        "type": "classification"
-        if kind in ("binary", "classification")
-        else "regression",
-        "range": list(value_range) if value_range else None,
-        "units": None,
-    }
-
-
-def _soft_head(entry: dict) -> dict:
-    """The surrogate's entry in a column's ``heads`` list."""
-    training = entry.get("training") or {}
-    return {
-        "role": "soft",
-        # The surrogate is fitted with squared error against the teacher's values (see
-        # CANONICAL_DEFAULTS), so it regresses whatever scale the teacher emits — including a
-        # probability. A future classification surrogate declares itself here and nowhere else.
-        "task": _task("regression", training.get("value_range")),
-        "source": {"kind": "teacher", "column": entry["name"]},
-        "training": {
-            "n": training.get("n_finite"),
-            "n_train": training.get("n_train"),
-            "n_val": training.get("n_val"),
-        },
-        "metrics": entry.get("metrics"),
-    }
-
-
-def _hard_head(entry: dict) -> dict:
-    """The hard-label head's entry in a column's ``heads`` list, read back off the run.
-
-    Its metrics are the alignment numbers, not classification scores: what decides whether this head is
-    trusted is how well its *calibrated* output reproduces the teacher's scale, and that is what
-    ``a_max`` is derived from.
-    """
-    meta_path = entry["dir"] / HARD_DIRNAME / HARD_META_NAME
-    meta = json.loads(meta_path.read_text()) if meta_path.exists() else {}
-    eval_path = entry["dir"] / HARD_DIRNAME / HARD_EVAL_NAME
-    evaluation = json.loads(eval_path.read_text()) if eval_path.exists() else {}
-    prepared = entry.get("hard_meta") or {}
-    gate = meta.get("tanimoto") or {}
-    calibration = evaluation.get("calibration") or {}
-
-    return {
-        "role": "hard",
-        # Taken from what the run resolved, not hardcoded: the labels decide, and a regression head is
-        # the case this field exists for.
-        "task": _task(prepared.get("task") or meta.get("task") or "binary"),
-        "source": {"kind": "measured", "column": prepared.get("source_column")},
-        "training": {"n": meta.get("n"), "n_positive": prepared.get("n_positive")},
-        "metrics": {
-            "r2_calibrated_soft": calibration.get("r2_calibrated_soft"),
-            "pearson_calibrated_soft": calibration.get("pearson_calibrated_soft"),
-            "spearman_h_soft": calibration.get("spearman_g_soft"),
-        },
-        "calibration": {"direction": calibration.get("direction")},
-        # The gate is a model in the graph like any other, so record what it is and the knees it ramps
-        # between. a_max of 0 means this head earned no weight and the column ships soft-only.
-        "gate": {
-            "signal": gate.get("signal"),
-            "a_max": gate.get("a_max"),
-            "t_lo": gate.get("sim_lo"),
-            "t_hi": gate.get("sim_hi"),
-        },
-        "provenance": {"lazyqsar_version": meta.get("lazyqsar_version")},
-    }
-
-
-def _bundle_metadata(
-    manifest: dict, plan: list, featurizer: dict, featurizer_class: str
-) -> dict:
-    """Everything a consumer needs, embedded in ``model.onnx`` so the file is the only input required.
-
-    Carries the featurizer (and the RDKit build it must run under), when the model was trained, the
-    reference library it was distilled from, and one entry per column describing that task.
-
-    The shape is N-ary throughout: ``columns`` is a list, and a one-endpoint model is simply its
-    one-element case. Within a column, every model that contributes is an entry in ``heads``, each
-    carrying its own ``task`` — so a classification surrogate, or a regression hard head, needs a
-    different ``type`` and no new keys. How the heads combine is *not* described here: the graph is the
-    only statement of that, and its single output per column is the only thing a consumer reads.
-
-    Nothing derivable is stored. ``n_columns``, the flat output list and "does this have a hard head"
-    are all properties of ``columns``, and duplicating them here is how they come to disagree with it.
-    """
-    import importlib.metadata
-
-    import rdkit
-
-    from olinda.data.fetch import MORGAN_FINGERPRINTS_FILENAME, MORGAN_FINGERPRINTS_URL
-
-    try:
-        olinda_version = importlib.metadata.version("olinda")
-    except Exception:
-        olinda_version = "unknown"
-
-    columns = []
-    for entry in plan:
-        heads = [_soft_head(entry)]
-        if entry["has_hard"]:
-            heads.append(_hard_head(entry))
-        columns.append(
-            {
-                "id": entry["id"],
-                "name": entry["name"],
-                "output": entry["output"],
-                # What the column itself emits. The blend maps the hard head onto the teacher's scale before
-                # mixing, so the column's task is the soft head's task whatever else contributes.
-                "task": heads[0]["task"],
-                "heads": heads,
-            }
-        )
-
-    split = manifest.get("split") or {}
-    return {
-        "schema": BUNDLE_SCHEMA,
-        "producer": "olinda",
-        "olinda_version": olinda_version,
-        "trained_at": datetime.now(timezone.utc).isoformat(timespec="seconds"),
-        "featurizer": {**featurizer, "rdkit_version": rdkit.__version__},
-        "featurizer_class": featurizer_class,
-        "reference_library": {
-            "name": MORGAN_FINGERPRINTS_FILENAME,
-            "url": MORGAN_FINGERPRINTS_URL,
-            **{
-                k: v
-                for k, v in (manifest.get("reference_library") or {}).items()
-                if k in ("n_rows", "dim")
-            },
-        },
-        "run": {
-            "backend": _run_backend(plan, manifest),
-            "created": manifest.get("created"),
-            **{k: split.get(k) for k in ("val_frac", "seed", "limit")},
-        },
-        "columns": columns,
-    }
-
-
-def _toposort(nodes: list, available: set) -> list:
-    """Order *nodes* so every input is produced before it is consumed.
-
-    ``add_prefix`` leaves each sub-model's nodes contiguous, and the Identity bridges that wire columns
-    together are appended afterwards, so the raw list is not in dependency order. Kahn's algorithm over
-    a producer map keeps this linear — a repeated-scan sort would be quadratic in the node count, which
-    grows with the number of columns.
-    """
-    producer: dict = {}
-    for i, nd in enumerate(nodes):
-        for out in nd.output:
-            producer[out] = i
-
-    indegree = [0] * len(nodes)
-    dependents: dict = {i: [] for i in range(len(nodes))}
-    for i, nd in enumerate(nodes):
-        for inp in nd.input:
-            if inp in available or inp not in producer:
-                continue
-            j = producer[inp]
-            if j != i:
-                dependents[j].append(i)
-                indegree[i] += 1
-
-    queue = [i for i, d in enumerate(indegree) if d == 0]
-    order: list = []
-    while queue:
-        i = queue.pop()
-        order.append(nodes[i])
-        for k in dependents[i]:
-            indegree[k] -= 1
-            if indegree[k] == 0:
-                queue.append(k)
-
-    if len(order) != len(nodes):
-        unresolved = [nodes[i].name for i, d in enumerate(indegree) if d > 0][:5]
-        raise ValueError(
-            f"fused graph has a dependency cycle or a missing producer near: {unresolved}"
-        )
-    return order
-
-
-def _parity_probe(plan: list) -> np.ndarray:
-    """Fingerprints to check the fused graph against the Python pipeline.
-
-    A handful of fixed molecules is not enough on its own: ``T`` may score them all
-    zero, in which case the blend collapses to the surrogate and the hard model, its calibrator and the
-    gate are compared against nothing — a cross-wired column would pass. A hard column's own labelled
-    compounds are at Tanimoto 1.0 from the labelled set, which is the top of the gate's ramp, so a few
-    of them are appended to force the whole blend to be exercised.
-    """
-    import h5py
-
-    from olinda.featurizer import MorganCountFeaturizer
-
-    blocks = [MorganCountFeaturizer().transform(_SAMPLE_SMILES).astype(np.float32)]
-    for entry in plan:
-        hard = entry["dir"] / HARD_H5_NAME
-        if not entry["has_hard"] or not hard.exists():
-            continue
-        with h5py.File(hard, "r") as f:
-            blocks.append(np.asarray(f["x"][:_PARITY_LABELLED], dtype=np.float32))
-    return np.concatenate(blocks, axis=0)
-
-
-def _run_backend(plan: list, manifest: dict) -> str | None:
-    """Which engine trained this run — recorded per column, never on the manifest itself."""
-    if manifest.get("backend"):
-        return manifest["backend"]
-    meta_path = plan[0]["dir"] / "train_meta.json" if plan else None
-    if meta_path and meta_path.exists():
-        return json.loads(meta_path.read_text()).get("backend")
-    return None
-
-
-def _assert_model_belongs_to(sm, entry: dict) -> None:
-    """Refuse to fuse a model that was trained for a different column.
-
-    Column directories are named positionally (``c0``, ``c1``, …), so re-preparing a run directory with
-    a different teacher file rebinds those names while leaving the previous run's artifacts in place.
-    Without this check the stale booster is fused under the new column's name, and the parity check
-    agrees because it validates against the same stale files.
-    """
-    trained_for = sm.metadata.get("column")
-    if trained_for is not None and trained_for != entry["name"]:
-        raise ValueError(
-            f"{entry['dir']} holds a model trained for column {trained_for!r}, but this run calls "
-            f"{entry['id']} {entry['name']!r}. The directory is stale — re-run `olinda learn-soft`, or "
-            "prepare into a clean directory."
-        )
-
-
-def _column_plan(model_dir: Path) -> tuple[dict, list]:
-    """The run manifest plus, per column, where its artifacts live and what its graph output is called.
-
-    Built once and shared by the fuse and the parity check, so both describe exactly the same sources.
-    """
-    from olinda import run as runlib
-
-    manifest = runlib.read_manifest(model_dir)
-    plan = []
-    seen: set[str] = set()
-    for col in manifest["columns"]:
-        col_dir = runlib.column_dir(model_dir, col["id"])
-        output = col["name"]
-        if output in seen:
-            raise ValueError(f"two columns would produce the output name {output!r}")
-        seen.add(output)
-        plan.append(
-            {
-                "id": col["id"],
-                "name": col["name"],
-                "output": output,
-                "dir": col_dir,
-                "has_hard": has_hard_head(col_dir),
-                "hard_meta": col.get("hard"),
-                "metrics": col.get("metrics"),
-                "training": {
-                    "n_finite": col.get("n_finite"),
-                    "n_train": col.get("n_train"),
-                    "n_val": col.get("n_val"),
-                    "value_range": col.get("value_range"),
-                },
-            }
-        )
-    if not plan:
-        raise ValueError(f"{model_dir} has no columns — run `olinda prepare` first")
-    return manifest, plan
-
-
-def _fuse(model_dir: Path):
-    """Assemble the fused ``model.onnx`` ModelProto: every column, sharing one input tensor.
-
-    Each column contributes an independent sub-pipeline under its own ``c{i}_`` prefix, so a run with
-    one column is simply the one-column case of a run with many. The graph exposes exactly one output
-    per column — its blended prediction, named after the column.
-    """
-    import onnx
-    from onnx import TensorProto, helper
-    from onnx.compose import add_prefix
-
-    from olinda.calibrate import IsotonicCalibrator
-    from olinda.featurizer import featurizer_from_meta
-    from olinda.models.bundle import StudentModel
-    from olinda.tanimoto import TanimotoRegressor
-
-    manifest, plan = _column_plan(model_dir)
-
-    nodes: list = []
-    inits: list = []
-    opset: dict = {"": _OPSET}
-
-    def collect(model, prefix: str) -> None:
-        m = add_prefix(model, prefix + "__")
-        nodes.extend(m.graph.node)
-        inits.extend(m.graph.initializer)
-        for op in m.opset_import:
-            opset[op.domain] = max(opset.get(op.domain, 0), op.version)
-
-    def ident(src: str, dst: str):
-        return helper.make_node("Identity", [src], [dst], name=f"br_{dst}")
-
-    def cast_d(src: str, dst: str):
-        return helper.make_node(
-            "Cast", [src], [dst], to=TensorProto.DOUBLE, name=f"br_{dst}"
-        )
-
-    # Tree ONNX outputs are (N,1); flatten to (N,) so they line up with the (N,) calibration/gate stages.
-    # Added once for the whole graph — a per-column copy would collide on name and fail the checker.
-    inits.append(helper.make_tensor("flat_shape", TensorProto.INT64, [1], [-1]))
-
-    def flat(src: str, dst: str):
-        return helper.make_node("Reshape", [src, "flat_shape"], [dst], name=f"fl_{dst}")
-
-    featurizer: dict = {}
-    featurizer_class = "MorganCountFeaturizer"
-    n_features = 2048
-    outputs: list[str] = []
-
-    for entry in plan:
-        p = (
-            entry["id"] + "_"
-        )  # every node and tensor of this column is namespaced by it
-        sm = StudentModel.load(entry["dir"], featurizer_factory=featurizer_from_meta)
-        _assert_model_belongs_to(sm, entry)
-        featurizer = sm.metadata.get("featurizer") or featurizer
-        featurizer_class = sm.metadata.get("featurizer_class", featurizer_class)
-        n_features = int(
-            featurizer.get("fp_size", sm.metadata.get("x_dim", n_features))
-        )
-        x_dim = int(sm.metadata.get("x_dim") or n_features)
-
-        # soft branch: input → soft_model → [soft correction] → surrogate (double)
-        collect(_soft_model_proto(sm, x_dim), f"{p}sm")
-        nodes.append(ident("input", f"{p}sm__input"))
-        nodes.append(flat(f"{p}sm__variable", f"{p}soft_raw"))
-        soft_raw = f"{p}soft_raw"
-        if (
-            sm.calibrator is not None
-        ):  # fitted during learn-soft and loaded from calibrator.json
-            collect(_isotonic_model(sm.calibrator, "in", "out"), f"{p}sc")
-            nodes.append(ident(soft_raw, f"{p}sc__in"))
-            surrogate_src = f"{p}sc__out"
-        else:
-            surrogate_src = soft_raw
-        nodes.append(cast_d(surrogate_src, f"{p}s"))
-
-        if entry["has_hard"]:
-            hard_root = entry["dir"] / HARD_DIRNAME
-            with open(hard_root / HARD_META_NAME) as fp:
-                hard_task = json.load(fp).get("task", "binary")
-
-            collect(
-                onnx.load(str(hard_root / HARD_MODEL_SUBDIR / "xgboost.onnx")), f"{p}hm"
-            )
-            nodes.append(ident("input", f"{p}hm__float_input"))
-            if (
-                hard_task == "regression"
-            ):  # seam: an H regressor exposes a single "variable" output
-                nodes.append(flat(f"{p}hm__variable", f"{p}g_reg"))
-                g_src = f"{p}g_reg"
-            else:  # classifier: take probabilities[:, 1]
-                collect(_prob1_model("p", "g"), f"{p}pr")
-                nodes.append(ident(f"{p}hm__probabilities", f"{p}pr__p"))
-                g_src = f"{p}pr__g"
-
-            gcal = IsotonicCalibrator.load(hard_root / H_TO_S_NAME)
-            collect(_isotonic_model(gcal, "in", "out"), f"{p}hc")
-            nodes.append(ident(g_src, f"{p}hc__in"))
-            nodes.append(ident(f"{p}hc__out", f"{p}h_s"))
-
-            clf = TanimotoRegressor.load(hard_root / TANIMOTO_DIRNAME)
-            collect(_tanimoto_model(clf, n_features, "input", "a"), f"{p}t")
-            nodes.append(ident("input", f"{p}t__input"))
-            nodes.append(ident(f"{p}t__a", f"{p}a"))
-
-            collect(_blender_model(), f"{p}bl")
-            nodes.append(ident(f"{p}s", f"{p}bl__soft"))
-            nodes.append(ident(f"{p}h_s", f"{p}bl__hard"))
-            nodes.append(ident(f"{p}a", f"{p}bl__a"))
-            nodes.append(ident(f"{p}bl__prediction", f"{p}prediction"))
-        else:
-            nodes.append(ident(f"{p}s", f"{p}prediction"))
-
-        nodes.append(ident(f"{p}prediction", entry["output"]))
-        outputs.append(entry["output"])
-
-    # One output per column and nothing else. ``S``, ``H_S`` and ``a`` remain as internal tensors — the
-    # blender consumes them — but they are working, not product: declaring them would invite callers to
-    # depend on a wiring we need to stay free to change.
-    out_vi = [
-        helper.make_tensor_value_info(n, TensorProto.DOUBLE, ["B"]) for n in outputs
-    ]
-    graph = helper.make_graph(
-        _toposort(nodes, {"input"} | {i.name for i in inits}),
-        "olinda_model",
-        [
-            helper.make_tensor_value_info(
-                "input", TensorProto.FLOAT, ["B", int(n_features)]
-            )
-        ],
-        out_vi,
-        initializer=inits,
-    )
-    model = helper.make_model(
-        graph, opset_imports=[helper.make_opsetid(d, v) for d, v in opset.items()]
-    )
-    md = _bundle_metadata(manifest, plan, featurizer, featurizer_class)
-
-    # Standard ONNX provenance, so the file identifies itself to any tool (Netron, hub tooling, the
-    # onnx CLI) without them having to know about the custom metadata key below.
-    model.producer_name = PRODUCER_NAME
-    model.producer_version = str(md.get("olinda_version", ""))
-    model.domain = "io.ersilia.olinda"
-    model.doc_string = (
-        f"Distilled model produced by olinda {md.get('olinda_version')} on {md.get('trained_at')}. "
-        f"Input: {int(n_features)}-d Morgan count fingerprint "
-        f"(RDKit {md['featurizer'].get('rdkit_version')}). Outputs: {', '.join(outputs)}."
-    )
-    entry_prop = model.metadata_props.add()
-    # json_safe first: metrics can carry NaN, which strict JSON parsers in other languages reject.
-    entry_prop.key, entry_prop.value = "olinda", json.dumps(json_safe(md))
-    return model, plan, outputs
-
-
-def build_bundle(model_dir: str | Path) -> dict:
-    """Fuse a run directory into ``model.onnx``, gated on numeric parity against the Python pipeline.
-
-    Every column is checked independently: the graph's output for that column must match a pure-Python
-    recomposition of exactly the same artifacts, so a mis-wired subgraph cannot slip through.
-    """
-    import onnxruntime as ort
-
-    from olinda.featurizer import featurizer_from_meta
-    from olinda.models.bundle import StudentModel
-
-    model_dir = Path(model_dir)
-    rule("olinda · export", style=STEP_COLORS["export"], right=cpath(model_dir))
-    model, plan, outputs = _fuse(model_dir)
-    echo(f"fusing {len(plan)} column(s) → model.onnx", "run")
-    _save(model, model_dir / MODEL_NAME)
-
-    echo("checking parity: model.onnx vs Python reference", "run")
-    ref: dict = {}
-    fp = _parity_probe(plan)
-    for entry in plan:
-        sm = StudentModel.load(entry["dir"], featurizer_factory=featurizer_from_meta)
-        raw = np.asarray(sm.predict(X=fp, calibrate=False)).ravel()
-        surrogate = (
-            np.asarray(sm.calibrator.transform(raw)).ravel()
-            if sm.calibrator is not None
-            else raw.astype(np.float64)
-        )
-        if not entry["has_hard"]:
-            ref[entry["output"]] = surrogate
-            continue
-
-        from lazyqsar.base.xgboost import BaseXGBArtifact
-
-        from olinda.calibrate import IsotonicCalibrator
-        from olinda.tanimoto import TanimotoRegressor
-
-        hard_root = entry["dir"] / HARD_DIRNAME
-        g = np.asarray(
-            BaseXGBArtifact.load(str(hard_root / HARD_MODEL_SUBDIR)).run(fp)
-        )[:, 1].astype(np.float64)
-        gsoft = np.asarray(
-            IsotonicCalibrator.load(hard_root / H_TO_S_NAME).transform(g)
-        ).ravel()
-        gate = TanimotoRegressor.load(hard_root / TANIMOTO_DIRNAME)
-        a = np.asarray(gate.weight(fp)).ravel()
-        # a == 0 everywhere has two very different causes. If the ceiling itself is zero the hard head did
-        # not earn any weight, the blend is off by design, and the fused output simply *is* the surrogate —
-        # there is nothing left to cross-check. If the ceiling is positive and the gate still never fires,
-        # the gate is broken and the hard branch would ship unverified.
-        if gate.a_max > 0 and not (a > 0).any():
-            raise RuntimeError(
-                f"parity probe for column {entry['name']!r} scored zero blend weight on every molecule, so "
-                "the blend collapses to the surrogate and the hard head would go unchecked. This should not "
-                "happen — the probe includes that column's own labelled compounds."
-            )
-        ref[entry["output"]] = (1.0 - a) * surrogate + a * gsoft
-
-    sess = ort.InferenceSession(
-        (model_dir / MODEL_NAME).read_bytes(), providers=["CPUExecutionProvider"]
-    )
-    got = {
-        o.name: np.asarray(v).ravel()
-        for o, v in zip(sess.get_outputs(), sess.run(None, {"input": fp}))
-    }
-    parity = {k: float(np.max(np.abs(got[k] - ref[k]))) for k in outputs}
-    worst = max(parity.values())
-    if worst > _PARITY_TOL:
-        offender = max(parity, key=parity.get)
-        raise RuntimeError(
-            f"model.onnx parity failed on column {offender!r}: max abs diff {worst:.2e} > {_PARITY_TOL:.0e}"
-        )
-
-    n_hard = sum(1 for e in plan if e["has_hard"])
-    size_b = (model_dir / MODEL_NAME).stat().st_size
-    size_txt = f"{size_b / 1e6:.1f} MB" if size_b >= 1e6 else f"{size_b / 1e3:.0f} KB"
-    summary_panel(
-        "olinda · export",
-        [
-            ("Columns", f"[bold]{len(plan)}[/] · {n_hard} with a hard head"),
-            ("Outputs", " · ".join(outputs)),
-            ("Parity (max)", f"[bold]{worst:.2e}[/] ≤ {_PARITY_TOL:.0e}"),
-            ("Size", f"[bold]{size_txt}[/]"),
-            ("Saved", f"[dim]{cpath(model_dir / MODEL_NAME)}[/]"),
-        ],
-        border_style=STEP_COLORS["export"],
-        icon="✓",
-    )
-    success(
-        f"fused model.onnx built and parity-checked → [dim]{cpath(model_dir / MODEL_NAME)}[/]"
-    )
-    return {"model": str(model_dir / MODEL_NAME), "columns": outputs, "parity": parity}
